修改：     CMakeLists.txt 	修改：     config/config_dual.json 	新文件：   config/test_io.json 	新文件：   doc/io_board.xml 	新文件：   examples/test_full_raw.c 	新文件：   examples/test_interactive.c 	新文件：   examples/test_io.c 	新文件：   examples/test_io_raw 	新文件：   examples/test_io_raw.c 	修改：     include/motor_api_internal.h 	修改：     include/motor_api_net.h 	修改：     include/motor_api_types.h 	新文件：   include/motor_driver.h 	新文件：   src/core/ma_core.c 	新文件：   src/core/ma_cycle.c 	新文件：   src/drivers/drv_cia402.c 	新文件：   src/drivers/drv_hans.c 	新文件：   src/drivers/drv_io.c 	删除：     src/motor_api.c 	新文件：   src/motor_api_facade.c 	修改：     src/motor_api_net.c 	重命名：   src/motor_api_config.c -> src/utils/ma_config.c 	重命名：   src/motor_api_common.c -> src/utils/ma_utils.c 	新文件：   test_full_raw 	新文件：   test_io 	新文件：   test_io_raw

+++ include/motor_api_net.h
@@ -28,7 +28,7 @@ EXTERNFUNC ma_status_t motor_api_set_command(struct motor_api_handle *handle,
                                              int step);
 
 EXTERNFUNC ma_status_t motor_api_set_axis_command(struct motor_api_handle *handle,
-                                                  int axis_idx,
+                                                  uint16_t axis_idx,
                                                   bool run,
                                                   int dir,
                                                   int step);
